+++ CHANGES.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## dev
 - remove `select_submission` - no longer needed with the tournament rule changes
+- make `upload` more robust
 
 ## v0.1.4 - 2022-02-04
 - added `set_comment` to set or update the comment associated with a submission